requirements.txt
streamlit 
pandas 
numpy
google-cloud-bigquery
db-dtypes